230714 자바수업

+++ JavaBasic/src/java00_javaReview/JavaReview_BaekJoon01.java
@@ -6,14 +6,16 @@ public class JavaReview_BaekJoon01 {
 	public static void main(String[] args) {
 
         Scanner sc = new Scanner(System.in);
-        int a,b,c;
-        a=sc.nextInt();
+        long a,b,c;
+        a=sc.nextLong();
         b=sc.nextInt();
         c=sc.nextInt();
         System.out.println((a+b)%c);
         System.out.println(((a%b)+(b+c))%c);
         System.out.println((a*b)%c);
         System.out.println(((a&c)*(b%c))%c);        
+        
+        
 	}
 }
